+++ packages/creamy_field/test/creamy_field_test.dart
@@ -1,5 +1,78 @@
+import 'package:creamy_field/creamy_field.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+const String initialText =
+    """Lorem ipsum is typically a corrupted version of De finibus bonorum et malorum;
+A first-century BC text by the Roman statesman and philosopher Cicero, with words altered, added, and removed to make it nonsensical, improper Latin.
+
+Versions of the Lorem ipsum text have been used in typesetting at least since the 1960s, when it was popularized by advertisements for Letraset transfer sheets. 
+Lorem ipsum was introduced to the digital world in the mid-1980s when Aldus employed it in graphic and word-processing templates for its desktop publishing program PageMaker. Other popular word processors including Pages and Microsoft Word have since adopted Lorem ipsum as well. """;
+
 void main() {
-  test('none', () {});
+  group('Tests for descriptions obtained from CreamyEditingController: ', () {
+    test('Text description test - 1', () {
+      const int _baseOffset = 0, _extentOffset = 11;
+      final CreamyEditingController textController1 =
+          CreamyEditingController.fromValue(
+        TextEditingValue(
+          text: initialText,
+          selection: const TextSelection(
+            baseOffset: _baseOffset,
+            extentOffset: _extentOffset,
+          ),
+        ),
+      );
+      expect(textController1.selectedText,
+          initialText.substring(_baseOffset, _extentOffset));
+      expect(textController1.totalLineCount, 5);
+      expect(textController1.atLine, 1);
+      expect(textController1.atColumn, 12);
+      expect(textController1.baseColumn, 1);
+      expect(textController1.extentColumn, 12);
+    });
+
+    test('Text description test - 2', () {
+      const int _baseOffset = 79, _extentOffset = 94;
+      final CreamyEditingController textController2 =
+          CreamyEditingController.fromValue(
+        TextEditingValue(
+          text: initialText,
+          selection: const TextSelection(
+            baseOffset: _baseOffset,
+            extentOffset: _extentOffset,
+          ),
+        ),
+      );
+      expect(textController2.selectedText,
+          initialText.substring(_baseOffset, _extentOffset));
+      expect(textController2.totalLineCount, 5);
+      expect(textController2.atLine, 2);
+      expect(textController2.atColumn, 16);
+      expect(textController2.baseColumn, 1);
+      expect(textController2.extentColumn, 16);
+    });
+
+    test('Text description test - 3', () {
+      int _baseOffset = initialText.length - 7,
+          _extentOffset = initialText.length;
+      final CreamyEditingController textController3 =
+          CreamyEditingController.fromValue(
+        TextEditingValue(
+          text: initialText,
+          selection: TextSelection(
+            baseOffset: _baseOffset,
+            extentOffset: _extentOffset,
+          ),
+        ),
+      );
+      expect(textController3.selectedText,
+          initialText.substring(_baseOffset, _extentOffset));
+      expect(textController3.totalLineCount, 5);
+      expect(textController3.atLine, 5);
+      expect(textController3.atColumn, 281);
+      expect(textController3.baseColumn, 274);
+      expect(textController3.extentColumn, 281);
+    });
+  });
 }